requirements.txt
python-telegram-bot[job-queue]==21.0.1
yt-dlp>=2024.03.10
python-dotenv==1.0.1
aiosqlite==0.20.0
